pyproject.toml
[project]
name = "agentic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "groq==0.30.0",
    "huggingface-hub>=1.8.0",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.3",
    "numpy>=2.2.6",
    "openai>=2.30.0",
    "qdrant-client>=1.17.1",
    "ragas>=0.4.3",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.3.0",
]
